requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
google-cloud-datastore==2.23.0